Move sample app source into `src` directory
