+++ example/src/components/Section.js
