+++ example/src/screens/SettingsScreen.js
